circular dependency fixed - sleep and coffee helped
